resize-api/requirements.txt
flask
Pillow
requests
redis